[#64] 🚚 move : MyPage components 폴더 생성 후 코드 이동
